Beginning restructure of the code.
In this resttucturing, the 1d simulation is completely modularized.  It
can be run stand-alone, as a script with input keyword arguments, or
can be imported as a module by some other routine (a fitting algorithm
to fit the tortuosity, for example).  The 1d_model file does three
simple things: initializes the model, runs the model, and if run as a
script, write output.

The user inputs are being moved to the cantera input yaml file.  The
initialization file calls this input, overwriting parametrs from this
file, as needed, based on user inputs.

Note that storing inputs in a yaml file introduces a new dependency
(ruamel.yaml), which can be installed via the conda-forge channel.